+++ client/src/components/KidProducts.js
@@ -308,7 +308,7 @@ class KidsProducts extends React.Component {
             let sizes = [];
             this.state.data.forEach(item => {
                 item.sizes.forEach(size => {
-                    sizes.push(parseInt(size));
+                    sizes.push(parseFloat(size));
                 })
             })
             const uniqueSizes = [...new Set(sizes)].sort((a, b) => a - b);
